BM1000: ConfigDxe: Add menu to configure power policy
When board is powered there are two options: boot immediately
or wait for power button press. User can select between "Off"
or "Auto" options. "Restore Last State" is not currently
supported but may be added if needed.
The Power Policy is stored in FRU and is checked by EC on start.